ShiftGraphView: refactor directory structor in model and view
