+++ src/shifts/shiftgraph/model/CMakeLists.txt
@@ -1,0 +1,8 @@
+set(MR_TIMETABLE_PLANNER_SOURCES
+  ${MR_TIMETABLE_PLANNER_SOURCES}
+
+  shifts/shiftgraph/model/shiftgraphscene.h
+  shifts/shiftgraph/model/shiftgraphscene.cpp
+
+  PARENT_SCOPE
+)
