+++ src/shifts/shiftgraph/view/CMakeLists.txt
@@ -1,0 +1,14 @@
+set(MR_TIMETABLE_PLANNER_SOURCES
+  ${MR_TIMETABLE_PLANNER_SOURCES}
+
+  shifts/shiftgraph/view/shiftgraphview.h
+  shifts/shiftgraph/view/shiftgraphview.cpp
+
+  shifts/shiftgraph/view/shiftgraphhourpanel.h
+  shifts/shiftgraph/view/shiftgraphhourpanel.cpp
+
+  shifts/shiftgraph/view/shiftgraphnameheader.h
+  shifts/shiftgraph/view/shiftgraphnameheader.cpp
+
+  PARENT_SCOPE
+)
